+++ xknx/remote_value/remote_value_1count.py
@@ -3,22 +3,32 @@ Module for managing an 1 count remote value.
 
 DPT 6.010.
 """
-from xknx.dpt import DPTArray, DPTValue1Count
+from typing import Optional, Union
+
+from xknx.dpt import DPTArray, DPTBinary, DPTValue1Count
 
 from .remote_value import RemoteValue
 
 
-class RemoteValue1Count(RemoteValue):
+class RemoteValue1Count(RemoteValue[DPTArray]):
     """Abstraction for remote value of KNX 6.010 (DPT_Value_1_Count)."""
 
-    def payload_valid(self, payload):
-        """Test if telegram payload may be parsed."""
-        return isinstance(payload, DPTArray) and len(payload.value) == 1
+    # pylint: disable=no-self-use
 
-    def to_knx(self, value):
+    def payload_valid(
+        self, payload: Optional[Union[DPTArray, DPTBinary]]
+    ) -> Optional[DPTArray]:
+        """Test if telegram payload may be parsed."""
+        return (
+            payload
+            if isinstance(payload, DPTArray) and len(payload.value) == 1
+            else None
+        )
+
+    def to_knx(self, value: int) -> DPTArray:
         """Convert value to payload."""
         return DPTArray(DPTValue1Count.to_knx(value))
 
-    def from_knx(self, payload):
+    def from_knx(self, payload: DPTArray) -> int:
         """Convert current payload to value."""
         return DPTValue1Count.from_knx(payload.value)